@@ -1,6 +1,6 @@
 module github.com/openstack-k8s-operators/lib-common/modules/ansible
 
-go 1.20
+go 1.21
 
 require gopkg.in/yaml.v3 v3.0.1
 
