@@ -23,7 +23,7 @@ import (
 	"sort"
 	"time"
 
-	"golang.org/x/exp/slices"
+	"slices"
 
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
