@@ -22,7 +22,7 @@ import (
 	"fmt"
 	"time"
 
-	"golang.org/x/exp/slices"
+	"slices"
 
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/util/validation/field"
